django_rest/requirements.txt
asgiref==3.8.1
Django==5.1.5
sqlparse==0.5.3
djangorestframework==3.15.2
django-debug-toolbar==4.4.6
Pillow==10.4.0
django-admin-interface==0.29.1
django-colorfield==0.11.0
drf-nested-routers==0.95.0
